@@ -13,12 +13,33 @@ class ProspectsController extends Controller {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
      */
     public function index()
     {
+        $prospects = Prospect::latest();
+        if(request('search')){
+            dd(request('search'));
+            $prospects->where('company', 'like', '%' . request('search'). '%');
+        }
         //Show all prospects and manage
-        return view('admin.prospects.index', ['prospects' => Prospect::latest()->paginate(20)]);
+        return view('admin.prospects.index', ['prospects' => $prospects->paginate(20)]);
+
+    }
+
+    public function all()
+    {
+        return Prospect::all();
+    }
+
+    public function filteredProspect()
+    {
+        $prospects = Prospect::latest();
+        if(request('search')){
+           $prospects = $prospects->where('company', 'like', '%' . request('search'). '%')->get();
+        }
+        //Show all prospects and manage
+        return $prospects;
 
     }
 
@@ -56,7 +77,7 @@ class ProspectsController extends Controller {
 
 
         // store prospect
-        return redirect()->route('admin.prospects.dashboard')->with('success', 'successfully created a new prospect');
+        return "Successfully created new prospect!";
 
     }
 
@@ -112,7 +133,7 @@ class ProspectsController extends Controller {
     {
 
         $prospect = Prospect::findOrFail($id)->delete();
-        if (\request()->wantsJson()) {
+        if (request()->wantsJson()) {
             return response()->json([
                 'alert_delete' => 'Selected query is deleted successfully.'
             ]);
@@ -150,11 +171,14 @@ class ProspectsController extends Controller {
 
 
 
-        return view('admin.prospects.search')->with('results', $results);
+        return $results;
 //        return view('admin.prospects.search', );
     }
 
-    public function get_prospect_data(){
+    public function get_prospect_data(): \Symfony\Component\HttpFoundation\BinaryFileResponse
+    {
+
         return Excel::download(new ProspectsExport, 'prospects.xlsx');
     }
 }
+
